Making sure ConfigurationFilePropertiesTest is run as part of the test suite
